requirements.txt
distribute==0.6.19
requests==0.7.4
wsgiref==0.1.2
python-dateutil==1.5